Add the ability to interact AWS
This takes the local-secure-file-storage app created by tyler19shea and changes the
file location to be an S3 bucket. The user must input the correct file to be downloaded
from the bucket as otherwise it will download a html element that is not useful. Further
error handling to be updated.